\name{chemCal-package}
\alias{chemCal-package}
\docType{package}
\title{
  Calibration functions for analytical chemistry
}
\description{
  See \url{../DESCRIPTION}
}
\details{
  There is a package vignette located in \url{../doc/chemCal.pdf}.
}
\author{
  Author and Maintainer: Johannes Ranke <jranke@uni-bremen.de>
}
\keyword{manip}
}
