% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemCal-package.R
\docType{data}
\name{din38402b1}
\alias{din38402b1}
\title{Nitrite calibration data}
\format{
A tibble containing 12 concentration levels with the respective
instrument response values.
}
\description{
Example dataset B.1 from DIN 38402 with concentrations in µg/L and the extinction
as response measured using continuous flow analysis (CFA) according to
ISO 13395.
}
\references{
DIN 38402-51:2017-05, Beuth Verlag, Berlin.
https://dx.doi.org/10.31030/2657448
}
\keyword{datasets}
