% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemCal-package.R
\docType{data}
\name{din38402b3}
\alias{din38402b3}
\title{Copper calibration data}
\format{
A tibble containing 13 concentration levels and the respective
instrument response values.
}
\description{
Example dataset B.3 from DIN 38402. Cu was measured according to ISO 11885,
using ICP-OES. The concentration are reported in mg/L and the response as
counts/s, describing the count of photons that are detected by the
photomultiplier detector of the device.
}
\references{
DIN 38402-51:2017-05, Beuth Verlag, Berlin.
https://dx.doi.org/10.31030/2657448
}
\keyword{datasets}
