% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemCal-package.R
\docType{data}
\name{din38402b6}
\alias{din38402b6}
\title{Carbamazepin calibration data}
\format{
A tibble containing 12 concentration levels and the respective
instrument response values.
}
\description{
Example dataset B.6 from DIN 38402 measured using LC-MS/MS. The
concentrations are reported in in µg/L and the response in arbitrary
units (AU).
}
\references{
DIN 38402-51:2017-05, Beuth Verlag, Berlin.
https://dx.doi.org/10.31030/2657448
}
\keyword{datasets}
