% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemCal-package.R
\docType{data}
\name{din38402c3}
\alias{din38402c3}
\title{Iron calibration data}
\format{
A tibble containing 10 concentration levels and the respective
response values.
}
\description{
Example dataset C.3 from DIN 38402 determined by ion chromatography.
Concentrations are reported in mg/L and the extinction as response.
}
\references{
DIN 38402-51:2017-05, Beuth Verlag, Berlin.
https://dx.doi.org/10.31030/2657448
}
\keyword{datasets}
