\name{draper}
\alias{draper}
\title{Regression example with repeated measurements}
\usage{data(draper)}
\references{Draper and Smith, Applied Regression Analysis (1981), p. 38}
\format{A dataframe with 24 observations on 2 variables}
\description{An example of a regression with multiple measurements per
factor level.}
\keyword{datasets}
