% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{linearity}
\alias{linearity}
\title{Assess the linearity of a calibration curve}
\usage{
linearity(x, y, method = NULL, range = 0.1)
}
\arguments{
\item{x}{numeric vector of independent values (usually concentrations)}

\item{y}{numeric vector of dependent values (usually the signal of the analytical device)}

\item{method}{character string. Supported methods are "point-to-point" and "curvature".}

\item{range}{numeric value describing the acceptable deviation from the median of the slopes or
the signal-to-concentration ratio}
}
\description{
A function to create diagnostic plots for the assessment of the linearity of
calibration data. The underlying methods can be found in DIN 32 402 (German Industrial Norm).
}
\examples{
# linearity(din32645$x, din32645$y, method = "point-to-point")
# linearity(din32645$x, din32645$y, method = "curvature", range = 0.2)

}
