\name{massart97ex1}
\docType{data}
\alias{massart97ex1}
\title{Calibration data from Massart et al. (1997), example 1}
\description{
  Sample dataset from p. 175 to test the package.
}
\usage{data(massart97ex1)}
\format{
  A dataframe containing 6 observations of x and y data.
}
\source{
  Massart, L.M, Vandenginste, B.G.M., Buydens, L.M.C., De Jong, S., Lewi, P.J.,
  Smeyers-Verbeke, J. (1997) Handbook of Chemometrics and Qualimetrics: Part A,
  Chapter 8.
}
\keyword{datasets}
