\name{pahCalibration}
\docType{data}
\alias{pahCalibration}
\title{Calibration data for HPLC measurement of 4 PAH}
\description{
  This dataset was produced during a course on trace analysis
  of organic contaminants. The data are far from perfect, but 
  good enough to serve as an example.
}
\usage{data(pahCalibration)}
\format{
  A dataframe containing the areas for the four polycyclic aromatic
  hydrocarbons (PAH) Acenaphthene, Phenanthrene, Anthracene and Pyrene.
  Each measurement of a standard solution makes up one row.
}
\source{
  \url{http://www.uft.uni-bremen.de/chemie}
}
\keyword{datasets}
