\name{pahMeasurements}
\docType{data}
\alias{pahMeasurements}
\title{Measurement data for HPLC measurement of 4 PAH}
\description{
  This dataset was produced during a course on trace analysis
  of organic contaminants. The data are far from perfect, but 
  good enough to serve as an example.
}
\usage{data(pahMeasurements)}
\format{
  A dataframe containing the areas for the four polycyclic aromatic
  hydrocarbons (PAH) Acenaphthene, Phenanthrene, Anthracene and Pyrene
  in the different samples.
}
\source{
  \url{http://www.uft.uni-bremen.de/chemie}
}
\keyword{datasets}
