% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery.R
\name{recovery}
\alias{recovery}
\title{Calculate the recovery}
\usage{
recovery(measured, expected)
}
\arguments{
\item{measured}{numeric value representing the measured concentration of the
analyte in the spiked test sample.}

\item{expected}{numeric value representing the expected concentration of the
analyte in the spiked test sample.}
}
\value{
a numeric value, being the relative recovery of the analyte.
}
\description{
Function to calculate the recovery of an analysis based on the ratio between
the measurement result of a sample that was spiked with a known amount of the
target analyte or a chemically similar substance.
}
