% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatability.R
\name{repeatability}
\alias{repeatability}
\title{Calculate the repeatability}
\usage{
repeatability(signal)
}
\arguments{
\item{signal}{numeric vector containing the signal values}
}
\description{
Function to calculate the repeatability
}
