% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_result.R
\name{calculate_result}
\alias{calculate_result}
\title{Calculate quantitative results}
\usage{
calculate_result(calibration, signal, df = 1, wf = 1, vf = 1)
}
\arguments{
\item{calibration}{object}

\item{signal}{numeric value}

\item{df}{numeric factor}

\item{wf}{numeric factor}

\item{vf}{numeric factor}
}
\value{
numeric value
}
\description{
Calculate the concentration or other things based on the signal of an
analytical device and a calibration.
}
