\name{din32645}
\docType{data}
\alias{din32645}
\title{Calibration data from DIN 32645}
\description{
  Sample dataset to test the package.
}
\usage{data(pahCalibration)}
\format{
  A dataframe containing 10 rows of x and y values.
}
\source{
  \url{http://www.uft.uni-bremen.de/chemie}
}
\keyword{datasets}
