% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{linearity}
\alias{linearity}
\title{Assess the linearity of a calibration curve}
\usage{
linearity(x, y, method = c("slope", "curvature"), tolerance = 0.1)
}
\arguments{
\item{x}{numeric vector of independent values (usually concentrations).}

\item{y}{numeric vector of dependent values (usually the signal of the
analytical device).}

\item{method}{character string. Supported methods are "slope" and
"curvature".}

\item{tolerance}{numeric value between 0 and 1, describing the acceptable
deviation from the median of the slopes or the signal-to-concentration
ratio. The default tolerance is 10\%.}
}
\value{
returns a diagnostic plot
}
\description{
A function to create diagnostic plots for the assessment of the linearity of
calibration data based on their point-to-point slope or the curvature.
The underlying methods follow ISO 84 66-1:2021 and DIN 32 402-51:2017
(German Industrial Norm).
}
\details{
The point-to-point slope method is based on the assumption that the slope
between two points should not vary greatly within the linear range.

The curvature method is similar to the point-to-point slope method. Here,
the ratio between the instrument signal and the concentration of the
calibration standard is assumed not to vary greatly within the linear range.

The use of the Mandel test is discouraged due to its limitations in the
identification of non-linear behaviour of calibration curves (Andrade and
Gomes-Carracedo, 2013).
}
\examples{
# Continuous Flow Analysis (CFA) data
data(din38402b1)

# Point-to-point slope plot
linearity(din38402b1$conc, din38402b1$ext, method = "slope")

# Curvature plot
linearity(din38402b1$conc, din38402b1$ext, method = "curvature")

}
\references{
ISO 8466-1:2021. Water quality — Calibration and evaluation of
analytical methods — Part 1: Linear calibration function

J. M. Andrade and M. P. Gomez-Carracedo (2013) Notes on the use of
Mandel's test to check for nonlinearity in laboratory calibrations.
Analytical Methods 5(5), 1145 - 1149.
}
\author{
Anıl Axel Tellbüscher
}
