\name{massart97ex3}
\docType{data}
\alias{massart97ex3}
\title{Calibration data from Massart et al. (1997), example 3}
\description{
  Sample dataset to test the package.
}
\usage{data(massart97ex3)}
\format{
  A dataframe containing 6 levels of x values with 5
  observations of y for each level.
}
\source{
  Massart, L.M, Vandenginste, B.G.M., Buydens, L.M.C., De Jong, S., Lewi, P.J., 
  Smeyers-Verbeke, J. (1997) Handbook of Chemometrics and Qualimetrics: Part A, p. 188
}
\keyword{datasets}
