% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemCal-package.R
\docType{data}
\name{rl95_cadmium}
\alias{rl95_cadmium}
\title{Cadmium concentrations measured by AAS as reported by Rocke and Lorenzato
(1995)}
\format{
A dataframe containing four replicate observations for each of the
six calibration standards.
}
\source{
Rocke, David M. und Lorenzato, Stefan (1995) A two-component model
for measurement error in analytical chemistry. Technometrics 37(2), 176-184.
}
\usage{
rl95_cadmium
}
\description{
Dataset reproduced from Table 1 in Rocke and Lorenzato (1995).
}
\keyword{datasets}
