% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemCal-package.R
\docType{data}
\name{rl95_toluene}
\alias{rl95_toluene}
\title{Toluene amounts measured by GC/MS as reported by Rocke and Lorenzato (1995)}
\format{
A dataframe containing four replicate observations for each of the
six calibration standards.
}
\source{
Rocke, David M. und Lorenzato, Stefan (1995) A two-component model
for measurement error in analytical chemistry. Technometrics 37(2), 176-184.
}
\usage{
rl95_toluene
}
\description{
Dataset reproduced from Table 4 in Rocke and Lorenzato (1995). The toluene
amount in the calibration samples is given in picograms per 100 µL.
Presumably this is the volume that was injected into the instrument.
}
\keyword{datasets}
